Move into MiniSQLMultisiteConnection, and add test for rollback
